+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,6 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
 import java.util.Queue;
 
 public class Deck {
