+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
